Refactor seems to be complete, still needs testing for insurance before merging to main
